requirements.txt
google-adk
google-genai
vertexai
gradio==5.34.2
